--- a/mapsforge-map/src/main/java/org/mapsforge/map/layer/renderer/DatabaseRenderer.java
+++ b/mapsforge-map/src/main/java/org/mapsforge/map/layer/renderer/DatabaseRenderer.java
@@ -3,6 +3,7 @@
  * Copyright 2014-2015 Ludwig M Brinckmann
  * Copyright 2014-2016 devemux86
  * Copyright 2017 usrusr
+ * Copyright 2024 Sublimis
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -83,7 +84,7 @@ public class DatabaseRenderer extends StandardRenderer {
      *
      * @param rendererJob the job that should be executed.
      */
-    public TileBitmap executeJob(RendererJob rendererJob) {
+    public synchronized TileBitmap executeJob(RendererJob rendererJob) {
         RenderContext renderContext = null;
         try {
             renderContext = new RenderContext(rendererJob, new CanvasRasterer(graphicFactory));