… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/AClasyHillShading.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/AThreadedHillShading.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/HgtCache.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/HiResStandardClasyHillShading.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/HillShadingUtils.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/MemoryCachingHgtReaderTileSource.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/ShadingAlgorithm.java
… mapsforge-map/src/main/java/org/mapsforge/map/layer/hills/StandardClasyHillShading.java
… mapsforge-map/src/main/java/org/mapsforge/map/rendertheme/renderinstruction/Hillshading.java
… mapsforge-map/src/test/java/org/mapsforge/map/layer/hills/RawShadingResultTest.java