@devemux86 2024-10-31
@Sublimis Thanks, quick question as there changes in the 2 Canvas classes:

There was a report about broken hillshading and high CPU at 0 latitude and 0 longitude:
- https://github.com/mapsforge/mapsforge/issues/1497

This was fixed with changes in the 2 Canvas classes:
- https://github.com/mapsforge/mapsforge/pull/1498

Have you tested the new code at 0 latitude and 0 longitude?