@Sublimis 2024-10-31
> @Sublimis Thanks, quick question as there are changes in the 2 Canvas classes:
> 
> There was a report about broken hillshading and high CPU at 0 latitude and 0 longitude:
> 
>     * [Hillshading empty regions and high CPU at 0 lat / lon #1497](https://github.com/mapsforge/mapsforge/issues/1497)
> 
> 
> This was fixed in the 2 Canvas classes:
> 
>     * [Fix hillshading at 0 lat / lon, fix #1497 #1498](https://github.com/mapsforge/mapsforge/pull/1498)
> 
> 
> Have you tested the new code at 0 latitude and 0 longitude?

Yup, no problems whatsoever. That was probably some manifestation of inefficiencies and/or poor synchronization that was taken care of in this PR. Screenshot:

![image](https://github.com/user-attachments/assets/307a57dd-7a3a-4570-a1e9-f1f854e9b13f)