@devemux86 2024-11-01
@Sublimis 

It produces pixelated images on older Android. Code on `master` works correctly.
Screenshot is with Mapsforge Samples on our min supported Android 5 (API 21):

<img src="https://github.com/user-attachments/assets/5d3cd8f3-bd13-443e-823c-50cf9992fe14" width="50%"/>