@Sublimis 2024-11-01
Easily fixed with a few lines of code in just `AndroidCanvas`. As it turns out, on Android 7 and older an `ALPHA_8` bitmap (like hill shadings) must be converted to `ARGB_8888` before upscaling, otherwise no filtering will be applied. Why it works in the current master is probably luck, because it uses a different implementation of the `android.graphics.Bitmap.createBitmap()` (inspection of the source code reveals that there are several, not to mention overrides), probably some undocumented magic happens there on old Androids.

Can we do a PR just to patch this up without repeating the previous big PR?
You can inspect the new code here: https://github.com/Sublimis/mapsforge/blob/hillshading-07/mapsforge-map-android/src/main/java/org/mapsforge/map/android/graphics/AndroidCanvas.java