@devemux86 2024-11-01
@Sublimis You posted an additional commit with changes in AndroidCanvas.
Are more changes needed?