@devemux86 2024-11-13
For now I reverted the `synchronized` in `DatabaseRenderer`
to avoid possible performance issues in normal map rendering.

---

For hillshading improvements, better only change the hillshading classes.

Changes in other classes should be discussed before to clarify their need
and be done separately in small pull requests per case.

This way it is easier to review and understand everything in small steps. :slightly_smiling_face: 

It looks like there are changes not related to hill shading (e.g. layers, labels).
Also `synchronized` and `volatile` in core classes need a good explanation as they _might_ affect performance.