@Sublimis 2024-11-13
We'll try to break down some of the changes in packages other than "layer.hills" as briefly as possible.

This code is called from multiple threads, so it's simply wrong:

https://github.com/mapsforge/mapsforge/blob/452dc3f07aa7dc496342fa9c67ef461a404dbb02/mapsforge-core/src/main/java/org/mapsforge/core/model/Tile.java#L222-L234

Even if it worked before, it's a disaster waiting to happen, it can cause bugs that are hard to find. The `getBoundingBox()` was made thread-safe in a similar fashion, as there are no guarantees that it won't be called from multiple threads (it probably is).

This code is also called from multiple threads:

https://github.com/mapsforge/mapsforge/blob/452dc3f07aa7dc496342fa9c67ef461a404dbb02/mapsforge-map/src/main/java/org/mapsforge/map/rendertheme/rule/RenderTheme.java#L80-L85
https://github.com/mapsforge/mapsforge/blob/452dc3f07aa7dc496342fa9c67ef461a404dbb02/mapsforge-map/src/main/java/org/mapsforge/map/rendertheme/rule/RenderTheme.java#L211-L213

This flag is also accessed from multiple threads:

https://github.com/mapsforge/mapsforge/blob/452dc3f07aa7dc496342fa9c67ef461a404dbb02/mapsforge-map/src/main/java/org/mapsforge/map/layer/renderer/RendererJob.java#L26C26-L26C26

This mutex is only used for hill shading code, in two places where needed to exclude concurrent access:

https://github.com/Sublimis/mapsforge/blob/05281fdf31b0246eb82a7b2cfcb6c6d544da5e8e/mapsforge-core/src/main/java/org/mapsforge/core/graphics/ABitmap.java#L19-L24

Learning from the situation with the mutex above, this field cannot be guaranteed to be modified from a single thread and should also be considered a disaster waiting to happen (in AWT there's a similar field):

https://github.com/mapsforge/mapsforge/blob/452dc3f07aa7dc496342fa9c67ef461a404dbb02/mapsforge-map-android/src/main/java/org/mapsforge/map/android/graphics/AndroidBitmap.java#L65

For this we merely introduced getter and setter for better encapsulation (it appears to be called only from one thread):

https://github.com/mapsforge/mapsforge/blob/452dc3f07aa7dc496342fa9c67ef461a404dbb02/mapsforge-map/src/main/java/org/mapsforge/map/rendertheme/RenderContext.java#L47C46-L47C46

`RenderContext.drawingLayers` was simply renamed to a singular form, since it actually holds a single layeR, not multiple layerS so it was very confusing 🥲 

Majority of the code mentioned is in fact somehow related to hill shading, except for these last two things. Tell us if we forgot something?

Also, the fields that have been made volatile are not contended, so this should be relevant: https://stackoverflow.com/questions/4633866/is-volatile-expensive

Thanks!