@devemux86 2024-11-13
> We'll try to break down some of the changes in packages other than "layer.hills" as briefly as possible.

Thanks! :+1: 

> This code is called from multiple threads

Why not adding `synchronized` to such methods?
(like we already do in [Layers](https://github.com/mapsforge/mapsforge/blob/master/mapsforge-map/src/main/java/org/mapsforge/map/layer/Layers.java) class)

> Majority of the code mentioned is in fact somehow related to hill shading, except for these last two things. Tell us if we forgot something?

Another example is changes in the constructor of `DatabaseRenderer` about the label layer.

There is no problem to change or improve other classes, but we have to do this separately,
in small steps and pull requests with good documentation about why the change is needed.

So we can easily review them in the future when we work again in these library areas. :slightly_smiling_face: